// src/frontend/webapp/modilist/src/services/swagger/api/models/line-item-feedback-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ModilistAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { LineItemSizeFeedback } from './line-item-size-feedback';
/**
 * 
 * @export
 * @interface LineItemFeedbackDTO
 */
export interface LineItemFeedbackDTO {
    /**
     * 
     * @type {number}
     * @memberof LineItemFeedbackDTO
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof LineItemFeedbackDTO
     */
    price?: number;
    /**
     * 
     * @type {LineItemSizeFeedback}
     * @memberof LineItemFeedbackDTO
     */
    size?: LineItemSizeFeedback;
    /**
     * 
     * @type {number}
     * @memberof LineItemFeedbackDTO
     */
    style?: number;
    /**
     * 
     * @type {number}
     * @memberof LineItemFeedbackDTO
     */
    fit?: number;
    /**
     * 
     * @type {number}
     * @memberof LineItemFeedbackDTO
     */
    color?: number;
    /**
     * 
     * @type {number}
     * @memberof LineItemFeedbackDTO
     */
    quality?: number;
    /**
     * 
     * @type {number}
     * @memberof LineItemFeedbackDTO
     */
    fabric?: number;
    /**
     * 
     * @type {number}
     * @memberof LineItemFeedbackDTO
     */
    pattern?: number;
    /**
     * 
     * @type {number}
     * @memberof LineItemFeedbackDTO
     */
    perfectMatch?: number;
    /**
     * 
     * @type {number}
     * @memberof LineItemFeedbackDTO
     */
    brand?: number;
    /**
     * 
     * @type {boolean}
     * @memberof LineItemFeedbackDTO
     */
    sendSimilarProducts?: boolean;
    /**
     * 
     * @type {string}
     * @memberof LineItemFeedbackDTO
     */
    additionalNotes?: string | null;
}